requirements.txt
streamlit==1.38.0
langchain-groq==0.1.0
langchain-core==0.1.52
langchain-community==0.2.6
python-dotenv==1.0.0
requests==2.32.3
yfinance==0.2.40
